+++ integration_test/tests/free_and_cancel_tests.rs
@@ -45,7 +45,7 @@ mod integration {
             query.push(0);
             assert_eq!(
                 SqlReturn::SUCCESS,
-                SQLPrepareW(stmt_handle, query.as_ptr(), SQL_NTS as i32),
+                SQLPrepareW(stmt_handle, query.as_ptr(), SQL_NTS),
                 "{}",
                 get_sql_diagnostics(HandleType::SQL_HANDLE_STMT, stmt_handle as Handle)
             );
